+++ backend/gateway/handler/strict.go
@@ -2,24 +2,29 @@ package handler
 
 import (
 	"code-connect/gateway"
-	"code-connect/pkg/ai"
 	"code-connect/problem"
 	"context"
 )
 
 type StrictHandler struct {
-	gptClient      ai.GPTClient
 	problemHandler *problem.Handler
 }
 
-func (s *StrictHandler) RequestProblem(ctx context.Context, request gateway.RequestProblemRequestObject) (gateway.RequestProblemResponseObject, error) {
+func (s *StrictHandler) GetMe(ctx context.Context, request gateway.GetMeRequestObject) (gateway.GetMeResponseObject, error) {
 	//TODO implement me
 	panic("implement me")
 }
 
+func NewStrictHandler(problemHandler *problem.Handler) *StrictHandler {
+	return &StrictHandler{problemHandler: problemHandler}
+}
+
+func (s *StrictHandler) RequestProblem(ctx context.Context, request gateway.RequestProblemRequestObject) (gateway.RequestProblemResponseObject, error) {
+	return s.problemHandler.RequestProblem(ctx, request)
+}
+
 func (s *StrictHandler) GetProblem(ctx context.Context, request gateway.GetProblemRequestObject) (gateway.GetProblemResponseObject, error) {
-	//TODO implement me
-	panic("implement me")
+	return s.problemHandler.GetProblem(ctx, request)
 }
 
 func (s *StrictHandler) GetRecords(ctx context.Context, request gateway.GetRecordsRequestObject) (gateway.GetRecordsResponseObject, error) {