+++ front/src/hook/Stepper.hook.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 
 const useStepper = () => {
     const navigate = useNavigate();
-
+    const [currentStep, setCurrentStep] = useState<StepType>("idle");
     const [StepperList, setStepperList] = useState<StepperListTypes>({
         difficult: {
             value: "난이도 체크",
@@ -22,20 +22,42 @@ const useStepper = () => {
             step: "idle",
         },
     });
+
+    const stepInit = () => {
+        setStepperList({
+            difficult: {
+                value: "난이도 체크",
+                step: "idle",
+            },
+            language: {
+                value: "언어 체크",
+                step: "idle",
+            },
+            api: {
+                value: "API 불러오는 중",
+                step: "idle",
+            },
+            comp: {
+                value: "문제 출제 완료",
+                step: "idle",
+            },
+        });
+    };
     const stepperStateChanger = (getStep: StepType) => {
+        setCurrentStep(getStep);
         if (getStep === "start") {
             StepperList.difficult.step = "loading";
-            const changeDefficult = StepperList.difficult;
+            const changeDifficult = StepperList.difficult;
             setStepperList((prevState) => {
-                return { ...prevState, difficult: changeDefficult };
+                return { ...prevState, difficult: changeDifficult };
             });
         } else if (getStep === "level") {
             StepperList.difficult.step = "complete";
             StepperList.language.step = "loading";
-            const changeDefficult = StepperList.difficult;
+            const changeDifficult = StepperList.difficult;
             const changeLanguage = StepperList.language;
             setStepperList((prevState) => {
-                return { ...prevState, difficult: changeDefficult, language: changeLanguage };
+                return { ...prevState, difficult: changeDifficult, language: changeLanguage };
             });
         } else if (getStep === "lang") {
             StepperList.language.step = "complete";
@@ -64,10 +86,41 @@ const useStepper = () => {
         }
         return true;
     };
+    useEffect(() => {
+        let count = 0;
+        let timer = 0;
+        timer = window.setInterval(() => {
+            count++;
+            console.log("count", count);
+            // if (selectStep === "clear") {
+            //     stepperStateChanger("clear");
+            //     return;
+            // }
+            if (count === 1) stepperStateChanger("start");
+            if (count === 2) stepperStateChanger("level");
+            if (count === 3) stepperStateChanger("lang");
+            if (count === 4) stepperStateChanger("api");
+            if (count === 4) clearInterval(timer);
+        }, 1000);
 
+        return () => {
+            stepInit();
+            clearInterval(timer);
+            console.log("stepper hook unmount");
+        };
+    }, []);
+    const endStep = () => {
+        // Stepper.tsx에서 실행
+        stepperStateChanger("clear");
+        setTimeout(() => {
+            stepperStateChanger("end");
+        }, 1000);
+    };
     return {
         StepperList,
         stepperStateChanger,
+        stepInit,
+        endStep,
     };
 };