+++ front/src/routes/Problem.tsx
@@ -1,31 +1,23 @@
-import React, { ReactNode, useState, useEffect, Suspense } from "react";
+import React from "react";
 
 import { motion } from "framer-motion";
 
-const ProblemComponent = React.lazy(() => import("../component/Problem/Problem.Component"));
+const ProblemNav = React.lazy(() => import("../component/Problem/Problem.Nav"));
+const ProblemBody = React.lazy(() => import("../component/Problem/Problem.Body"));
 
 function Problem() {
     return (
         <motion.div
+            className="h-full w-full"
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
             exit={{ opacity: 0 }}
             transition={{ duration: 0.5 }}
         >
-            <ProblemComponent />
+            <ProblemNav />
+            <ProblemBody />
         </motion.div>
     );
-    // return (
-    //     <motion.div
-    //         className=""
-    //         initial={{ opacity: 0 }}
-    //         animate={{ opacity: 1 }}
-    //         exit={{ opacity: 0 }}
-    //         transition={{ duration: 0.5 }}
-    //     >
-    //         <ProblemComponent />
-    //     </motion.div>
-    // );
 }
 
 export default Problem;